Besp03008 update and create ib ob pr

+++ src/main/java/org/example/wms_be/entity/inbound/PurchaseDetailsIb.java
@@ -2,6 +2,8 @@ package org.example.wms_be.entity.inbound;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class PurchaseDetailsIb {
@@ -11,9 +13,9 @@ public class PurchaseDetailsIb {
     private String  maPO;
     private Integer  sysIdKhachHang;
     private Integer  sysIdSanPham;
-    private Double  soLuong;
-    private Double  gia;
-    private Double  tongChiPhi;
+    private Double soLuong;
+    private BigDecimal  gia;
+    private BigDecimal  tongChiPhi;
     private String  ngayNhapDuKien;
     private String maLoHang;
     private String tenSanPham;
