+++ src/main/java/org/example/wms_be/entity/outbound/PurchaseRequestDetailsOb.java
@@ -4,6 +4,8 @@ package org.example.wms_be.entity.outbound;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class PurchaseRequestDetailsOb {
@@ -15,7 +17,7 @@ public class PurchaseRequestDetailsOb {
     private Integer  sysIdSanPham;
     private String tenSanPham;
     private Double  soLuong;
-    private Double  gia;
-    private Double  tongChiPhi;
+    private BigDecimal gia;
+    private BigDecimal  tongChiPhi;
     private String  ngayXuatDuKien;
 }
